Serve a 422 when rummager returns 422
At present when requests fail Rummager's validation, Rummager will return a
422 (unprocessable) response.  Finder-frontend receives this as a
GdsApi::HTTPUnprocessableEntity exception, which results in a 500 error.

If we return a 422 then it makes it clear to the user (usually a script exploring
the site for vulnerabilities) that we're handling the request appropriately
and that they need to do something else.

This also reduces the 500 count for alerting purposes, which at times is very
noisy.  It's hard to tell the difference between legitimate server issues and
vexatious use.