Enable faceting by World Location
World Locations will now be fetched from Whitehall (if possible)
for use in faceting. Only world location slugs are returned
by rummager so we need to get additional world location
data from Whitehall. This data is cached in a Registry.

I added a World Locations Registry to lib, similar to the
registries we have in Rummager. Initially, I had planned to
add this world location registry to rummager. Having the
registry further downstream (it would only be used by
finder-frontend) ensures there's no dependency on
Whitehall in rummager.

This unmerged PR demonstrates how this could have been
implemented in Rummager:
https://github.com/alphagov/rummager/pull/1318/commits/d54d0a6e3f22148ea645cf8c34333c1174fd9fc2.

By default, the world locations finder will not be
populated in development mode. It relies on Whitehall
being up, so if whitehall-admin happens to be up, or if
you start finder frontend with the --live flag, the world
locations registry will be populated with data.

A longer term solution would be to bring world locations
back into rummager. This is being considered but it may
be some time before this is worked on.