Add TopicTaxonomyRegistry
This new Registry, like the WorldLocationsRegistry, is a
caching wrapper around an external API for the purpose
of providing a facet on finders.

This registry is required as the aggregates for taxon
filters returned by rummager do not provide the required
detail for faceting (titles are missing, only content_ids
are provided).

A subsequent commit will use the registry in a finder.