Add Topic faceting to finders
This makes it possible for end users to filter by top-level
taxons (referred to as topics). Making use of the new taxon
registry added in a previous commit.

This uses the select component [1] from
govuk-publishing-components.

Further work is needed to make it possible to filter by
second-level taxons, and to ensure that the brexit
checkbox works well with this new filter.

[1] https://govuk-publishing-components.herokuapp.com/component-guide/select