Change email signup link when saving state
This means that the email signup link will get changed when the
page loads as well as when the facets are altered. Currently when
a user visits the finder with a URL that preselects facets
(i.e. from the Q&A) the email signup link doesn't keep the same
facets selected.