Sub-Topic Faceting
This makes it possible to filter by sub-topic in addition to topic.
It makes the following changes:
- JS changes to handle taxonomy faceting logic
- facets may have multiple keys, so we can have multiple URI keys for the
same filter. This is useful if we want two input fields to point to the
same search api field. E.g. top_level_taxon = x, second_level_taxon = y in
the url, and to send a different field to rummager like part_of_taxonomy=[x,y].

This also adds a gem, dalli, for in memory cacheing during tests.