Omit search box when there are no results
We want the search box to be visible if there:
 - are results
 - are keywords specified
 - are other facets present that may have filtered the results to zero

We're trying to reduce the number of times the search box is present when
it is useless.  A drawback is that we don't want it to keep appearing and
disappearing without an obvious reason for it.  For this reason, we're being
a little cautious about the scenarios in which this occurs.

I suspect further user testing might allow us to expand on this.