Allow content store document type and supertypes
Previously if a supertype was specified, it would overwrite any content
store document types that were present.  This means that you couldn't do
ad hoc combinations of supertypes and document types.

We now have a need to do this so that we can include a specific format in
a finder.