Fix Remove Filter function for Autocomplete items
There is an issue where filters selected from an autocomplete option
select [1] do not get removed when the user clicks on the remove icon.

This will fix that. It is caused by the weird way that autocomplete
works, in that there are two input/selects that a user changes when
they use the autocomplete component.