Fix facet tags for accessible-autocomplete
This commit fixes the facet tags for the accessible-autocomplete component so that clicking on the remove button has the desired behaviour of removing this filter from the search.

In order to achieve this, it is necessary to determine if the input we're dealing with has created by the accessible-autocomplete component. The `accessible-autocomplete` does not set the `name` attribute of the generated `<input>` element, so this is not helpful to us.

Instead, an `accessible-autocomplete` will enhance a `select` element and rename that element to be `<ID>__listbox`, where `<ID>` is the original `id` attribute on the `select` element. Therefore we can look for the presence of a `select` with an id == `<FACET_VALUE>__listbox` to determine if we're dealing with an autocomplete. If this is the case, we can get the input without needing to `find` anything further inside the element, giving us the appropriate element type and input type, before entirely replacing the current value of the autocomplete with a blank string, removing the filter from the search.

Solo: @karlbaker02