Fix bug with checkbox facet and facet tags
This commit fixes a bug which causes the facet tags to disappear if the following are performed in order (more cases are avaiable, this is an example):

- Choose a topic
- Filter by content related to Brexit only by clicking on checkbox
- Remove content related to Brexit only filter by clicking on checkbox

Once here, the facet tags and the number of results will disappear, even though a) the topic facet tag is still present and b) there are only a subset of results available.

The checkbox facet partial was using `controls` - this should instead be `aria_controls_id`, as is the case with all of the other facet partials.

Solo: @karlbaker02