Allow filtering on document_type and supergroup
This will allow users to filter by document type and supergroup.

Supergroup filters are to be deprecated. Internally, finder-frontend
has a class TranslateContentPurposeFields that transforms supergroup
filters (content_purpose_supergroup) to document type filters
(content_store_document_type).

The problem is that, if you have a document_type facet/filter, then
this can only expand the search. For example, if you are viewing the
news and communications finder, this will be filtered by
the news_and_communications supergroup (this gets turned into a
document_type filter of document_type[]=news_story,article,... etc).
If you wanted to just view documents of type press release on that
finder, this wouldnt be possible.

This code enables one to filter by supergroup OR by a provided set
of document types. The supergroup will be the default unless the
document types are provided. The expected use case is filtering
down supergroups by subgroupings of document types.

An example of this is included in this commit, where I make
the research_and_statistics finder filterable by document type
sets of Published/Upcoming/All statuses of stats.