Add acceptance tests for live sorting
This configures the Cucumber acceptance tests to use the Chrome browser to assert the correct live search finder behavior. Previously this was not being tested as the tests were using the Rack Test driver.

This builds on top of #805 which updated Slimmer's test templates to include the necessary JavaScript feature detection hooks needed to enable the JavaScript live search behaviour.

Part of https://trello.com/c/VENEFcz5