Add SignupUrlHiddenParams presenter
This presenter prepares the parameters for the email signup button
on the email signup page.

This class looks at the email signup content item to see what
facets are permitted, and maps the correct keys and values
from those provided from the finder link to the signup page.

When a user clicks the Email signup button, these hidden params
are sent to EmailAlertSubscriptionsController#create and are
processed into a subscriber_list creation request before we
redirect the user to the email-api-frontend.