Add ValidateQuery
This enables us to validate that parameters are acceptable to
rummager. This is useful in the context of dynamic filters
being passed to finders and email subscriptions. These filters
can be provided to finder-frontend by users. Aside from the
current whitelist approach (specifying permitted filter keys
in finder content items), it is also useful to test whether
the query will throw errors before saving it as an email
subscription. In this case, it is possible for the user
to provide filters that have not been whitelisted.
This is a problem because the filters will be saved and used
by email-alert-api. So validating the query is useful
as it ensures that the user is giving us something that
wont throw errors later on.