Dynamic facets in email subscription titles
This adds dynamic facets to an email subscription title
(displayed in email-alert-frontend).

Previously the approach to generating a title for a
subscription was to list all the choices for a facet
such as _case type_ in the email signup content item.
The title would be generated using the values from the
item and the params from the users choices.

This isnt workable with dynamic facets like organisations
and people, where there could potentially be thousands of
options which are constantly changing.

For finder facets we chose to define registries for
topics and world locations; we will probably do something
similar for organisations and people. Alternatively we could
fetch the complete aggregate list for each data type
from rummager.

In the meantime, we whitelist the dynamic filters like
people and organisations and display the count of the
dynamic filters rather than the beautified value (we only
have the ids and slugs to send to email-alert-api).

See the spec for a description of what this does in
practice.