Send email signup content item default attributes to email-api
This allows us to send default filters from a content item
for email signup to the email-alert-api.

For example, if a user subscribes to an email on
the news and communications finder, this has hidden
filters such as on content_purpose_supergroup.

The user will expect that this filter is passed
to the email subscription, and that it would be hidden
from them.

Additionally, this adds validation of email subscriptions
before we send them to the email-alert-api, making use
of ValidateQuery.