Adjust checkbox facet properties
The checkbox facet adds a few new schema properties which we can avoid if we
reuse some of the existing ones a bit more judiciously.

I've removed label, checkbox_label and value from the required properties. I've
defaulted value to true (and we can add that property if we really need it).
The others seem to be OK if we just rejig how name and short_name are used.

https://trello.com/c/oE1sdBdt/278-add-ability-to-publish-news-and-communications-finder-from-rummager