Allow Atom feeds on more things
When the default order is set to -public_timestamp, this matches the behaviour
of an atom feed.

I've also added some tests around this behaviour so that we can be more
confident about changing it in future.  I'm not sure it's quite right to
expect the atom feed to behave exactly like search results in a finder.