Enable ALL for email-subscription tags
Previously, it was only possible to subscribe to taxons
in a non-conjunctive way (brexit OR education). So one
would receive emails whenever there was an email related
to brexit regardless of whether it was related to education.

This introduces conjunctive queries so we can subscribe to
email related to brexit AND education.