Add registry to power the org breadcrumb
At present we're fetching a content item to check whether the supplied parent
path is a valid organisation.  This can be quite slow because of the latency to
content store at present, and may even cause the page to time out altogether
which I would say is not acceptable.

Because of that, I'm adding another registry to get all the slugs and titles of
the organisations that we support.  We can look the org slug up against that and
only have to refresh it every so often.  The query for all orgs takes about
0.09 seconds for Rummager to process.