Cache content store items for 5 minutes
This will help level out the latency issues for most users. We're going to
be adding a lot of traffic to finder frontend, so we should keep it as fast
as we can.

We need the content item to render the finder, so it's OK to error if the
request times out.

Content items used to be cached everywhere until https://github.com/alphagov/gds-api-adapters/pull/848.
This was a good idea for most content, but less suited to static things
like finders.

The QA controller also does this: https://github.com/alphagov/finder-frontend/blob/master/app/controllers/qa_controller.rb#L27

Finder content items rarely change, so this should be OK.