Fix tracking bug where showing incorrect result count
when the pageview event is fired after changing filter/sort.
dimension 27 should contain the number of results.

Currently it shows the number of results for the previous page view.

This is because the event pageview event is fired before
the results are received.

Moving triggering the pageview event into the `done` handler for updateResults.