Fix tracking on facet tag removal.
We had a race condition where we could not tell if the `facetTagRemoved`
GA event would be sent or if the data would be reloaded first.

We need the `facetTagRemoved` to fire first because we need it to contain the
number of items displayed when the user clicked on the link.

In an attempt to stop the form submission, I added an `e.stopPropogation()`.

It didn't stop the form from being submitted but `seems` to have allowed the
GA event to happen before the data gets re-loaded, which is the behaviour we are
looking for.

I say `seems` here because it's a race condition the only happened rarely.