Fix issue with Capybara tests failing on removing facet tags
Previously, Capybara was failing on tests involving removing facet tags because the `RemoveFilter` JavaScript module was not being associated. It seems to be the case that although the module works fine in both development and production, the test setup with Capybara meant that the module's `start` method would never be called.

The suspected reason for this is that Capybara tests are not using Slimmer/Static. After consulting with @alex-ju, a call to `require govuk_publishing_components/dependencies` in the application.js file was added, as this manually calls `GOVUK.modules.start()` and enables `RemoveFilter` to be seen in the tests -  [further info here](https://github.com/alphagov/govuk_publishing_components/blob/c0f608d18eb91fc834a8668930953bcd8a94b48a/docs/install-and-use.md#javascript)

Solo: @karlbaker02