Add PeopleRegistry
This new registry will enable us to provide values
when provided with people slugs (either in email
alert title builder or by finder facets).

We have similar registries for e.g. organisations so
this is a continuation of that pattern.

https://trello.com/c/J6ZHmdgi/349-improve-email-signups-description

A followup commit will make use of this registry
in email alert title builder.

Authors: Fli.* & Bill F