Add more detail to subscriber list title for dynamic filters.
Before:

"News with 3 people and 2 organisations"

After:

"News with people of Harry Potter, Ron Weasley, and 1 other, and organisations of HM Treasury and Cabinet Office"

It is possible to display the actual values rather than the counts because we now
have registries - cached collections of people, organisations, world locations, etc. -
that we can use to lookup a beautified value for a given slug.

Authors: Felicity and Bill

https://trello.com/c/J6ZHmdgi/349-improve-email-signups-description?menu=filter&filter=label:Finders