Show all values for People, Organisations and World Locations regardless of applied filters
This commit updates finders to show all available _People_, _Organisations_ and _World Locations_ values in dynamic facets, regardless of whether these values are applicable to the currently selected facets.

The reason behind this is to ensure that we don't give the appearance of les options being available for these facets than there actually are, and also to ensure that we don't incorrectly constrain a user's search as they select and de-select filters - something which can happen right now, as the facets aren't reloaded on-the-fly with JS enabled but instead only on page load/reload.

Further information can be found at https://trello.com/c/Bc9UKt5W.

Solo: @karlbaker02