Add Cucumber test to ensure all values for dynamic facets values are shown after reload
This commit adds a new Cucumber test to ensure that all dynamic facet values are still shown once a user has selected some combination of facets and then reloads the page. This is then equivalent to what a new user would experience if they had been given a link to the finder, and then ensures that they can still see all of the values for the dynamic facets, even if some of those values aren't related to the current applied filters.