Put closed orgs at the bottom of the facet
We need to show closed organisations as they still have content tagged to them.
To help users navigate a long list, we're going to move them to the bottom.  At
present they appear in a clump as the orgs list is sorted by title, and closed
orgs start with "Closed organisation: ".

We also want to use a case insensitive sort because we have one org that starts
with a lower case "j".

https://trello.com/c/leCxXdgK/367-closed-orgs-should-appear-at-the-bottom-of-the-org-autocomplete-facet