Remove keyboard listeners from option select
The option select used to just use a `<div>` for the select head, which meant that we had to take care of keypresses manually [1].

Shortly after, the code was updated to use a button for the head [2]. By using a button we get keyboard access 'for free' – you can trigger a button using either space or the enter key – but the logic to handle keypresses wasn't removed.

We can now remove it, because it's redundant.

[1]: https://github.com/alphagov/static/commit/8f42ee21f9c88efd9b8bfcbb8e6f026830a58748
[2]: https://github.com/alphagov/static/commit/04225fdde1a2f9cc59a9f0453d6a175693f1217e