Remove redundant focus handler from option select
When the option select was initially introduced, when option selects were 'closed' they were actually just hidden off screen [1].

This meant that it was possible to tab to the checkboxes within it, without having 'opened' the option select.

We therefore have a event handler set up which will attempt to open the option select when you focus a checkbox within it.

However, we now 'close' an option select by setting `display: none`. As far as I am aware, this takes the checkboxes out of the tab order, which means it's impossible (?) to focus them without first opening the option select.

This means that this code is redundant.

[1]: https://github.com/alphagov/static/blob/8f42ee21f9c88efd9b8bfcbb8e6f026830a58748/app/assets/stylesheets/govuk-component/_option-select.scss#L77-L82
[2]: https://github.com/alphagov/finder-frontend/blob/master/app/assets/stylesheets/components/_option-select.scss#L100-L102