Add option lookup to option-select facet
This commit introduces a new option within the `option-select` facet to enable individual options as rendered by a checkbox to map to multiple values. This is done via a new property defined in the facet configuration called `option_lookup`, which defines the key-to-values relationship. In the `allowed_values`, the value for each of the options then corresponds to the key you set for it in `option_lookup`.

For example, the following JSON is defined for the `content_store_document_type` facet in the Policy finder:

```
{
  "key": "content_store_document_type",
  "name": "Document type",
  "preposition": "with document type",
  "type":"text",
  "display_as_result_metadata": true,
  "filterable": true,
  "option_lookup": {
    "policy_papers": ["impact_assessment", "case_study", "policy_paper"],
    "open_consultations": ["open_consultation"],
    "closed_consultations": ["closed_consultation", "consultation_outcome"]
  },
  "allowed_values": [
    {
      "label": "Policy papers",
      "value": "policy_papers"
    },
    {
      "label": "Consultations (open)",
      "value": "open_consultations"
    },
    {
      "label": "Consultation (closed)",
      "value": "closed_consultations"
    }
  ]
}
```

The options presented to the user as checkboxes in this component will have labels of `Policy papers`, `Consultations (open)` and `Consultations (closed)`, which map to values `policy_paper`, `open_consultations` and `closed_consultations` respectively. Now, because each of these values needs to map to multiple document types, the `option_lookup` property is set, and this maps each of these values (`policy_paper`, `open_consultations` and `closed_consultations`) to the corresponding document types necessary for the finder to search.

One of the main reasons for creating a pre-defined lookup within this component is to be able to filter by multiple values without needing to encode a JSON array within the value - if we had used this approach we'd need to parse the JSON on every request, which could have security implications. Instead, by encoding these options in the facet's content item, we remove the need to parse any JSON, in addition to reducing the payload sent by the user.