Remove block-label class
In eee8cfb we updated finder-frontend to use the checkboxes from govuk_publishing_components.

The previous checkboxes appear to be the only place where the block-label class was used, so these styles are no longer required.

Verified by searching the codebase for the string ‘block-label’ – the only other instance was for instantiating the SelectionButton object in application.js, which is being removed in a separate PR.