Allow users to sort results by relevance
In PR #928, functionality was added to change the default ordering to
"relevance" if users do a keyword search, otherwise the default sort
order remains unchanged.
That functionality is being added to the business readiness finder.

Note: As "Relevance" is an optional sort option, it needs to be added to
the middle of the list of sort options in the finder content item.
That's because the javascript that does the reordering fails if it can't
find an "adjacent" option i.e. the option after it in the dropdown list.

Pair: @mahmudhussain @leenagupte