Add tests to check default filtering
The default filter option on the business readiness finder changes
depending on the search. If the user only searches by facet, the
default order should remain "Topic", however if the user searches by
keyword, the default ordering should be changed to "Relevance".

These tests make sure that the business readiness finder behaves as
expected, even if the default behaviour of the other finders change.