Introduce new filter type `content_id`
This filter type assumes that each item the `allowed_values`
list contains a `content_id`

This `content_id` is then used for the query to rummager, instead
of the `value` that is used for the old metadata style
queries.

This is a step towards replacing metadata tagging with entries
in the `links` hash.