Allow option_select component to show/hide filter field
By default the option_select component will not display the filter field.
The reason for this is that we wanted to control which option_select facet
allowed users to filter the checkboxes. There maybe a scenario where the
facet only has a few checkboxes and therefore it wouldn't make sense to
filter the options out (For example Policy finders doc type filter which
only has 3 options)