fixes "disabled" Relevance sort option
By default "Relevance" sort option is disabled when JS has been disabled.
When the livesearch initialises it remembers the state of "Relevance" sort
option and then inserts the option back into the sort list in its original
state. This means that the option is disabled and users cannot reselect it
should they want to.