Refactor step definitions to prevent breaking others.
At present the `I fill in some keywords` step stubs all
rummager requests with the business finder results.

This causes unrelated scenarios to break when I change
the format of the business finder response to something
incompatible with those finders.

This commit refactors the stubbing of results for
business finder scenarios into the
`I view the business readiness finder` step.