Get content id of email signup page
The way email signups work for the business readiness finder is
changing. This is because the finder works differently to other finders.
Rather than narrowing the results, the finder expands them, doing an
"or" query rather than an "and". That is not how email-alert-api works.
It creates an "and" query from the facets it's passed.

This means that most users subscribed to the business readiness finders
were not receiving updates.

Passing the content id to EmailAlertSignupAPI so we can use it to
determine whether we need to overwrite the tags passed to
email-alert-api.