Overwrite tags for email signup
The way email signups work for the business readiness finder is
changing. This is because the finder works differently to other finders.
Rather than narrowing the results, the finder expands them, doing an
"or" query rather than an "and". That is not how email-alert-api works.
It creates an "and" query from the facets it's passed.

This means that most users subscribed to the business readiness finders
were not receiving updates.

We need to pass the correct tags and title to email-alert-api to create
the right subscription.

We know we need to pass:

```
{ "appear_in_find_eu_exit_guidance_business_finder" => { "any" => %W(yes) } }
```

in the tags.

We tried various options to add this as a `filter` in the email-signup
content item to no avail:

`email_filter_by: appear_in_find_eu_exit_guidance_business_finder`
This returned `all-government-publishing-updates` as the subscription
topic_id

`filter: "appear_in_find_eu_exit_guidance_business_finder"`
(both with and without the subscription_title_prefix whitespace)
We had to add it to the content schemas. This also returned
`all-government-publishing-updates` as the subscription topic_id

```facet_id: appear_in_find_eu_exit_guidance_business_finder
filter_key: "appear_in_find_eu_exit_guidance_business_finder"
filter_value: "yes"
facet_name: appear_in_find_eu_exit_guidance_business_finder```
This also returned `all-government-publishing-updates` as the
subscription topic_id