Check business readiness tags are passed to email
This assumes that if "appear_in_find_eu_exit_guidance_business_finder" =
"yes" is passed to the filtering options are passed correctly to
EmailAlertSignupAPI the tags will be generated correctly.

At this point we don't know how the tag appears in the content_item.

The title builder had to be updated so that a trailing "with" wasn't
appended to the email title.