Hide checkbox
We don't want users to see the
“appear_in_find_eu_exit_guidance_business_finder” checkbox because they
should not be able to change the value. It should always be "yes".

Another option is to add “appear_in_find_eu_exit_guidance_business_finder”
as an allowed "filter" in the email-signup content-item. This requires
a schema change, but it means that we would not need to add the extra
code to hide the checkbox