Radio facet & filter will use provided default value if unset
This ensures that the default in allowed_values for a
radio button facet is used if provided and no value
is set for the radio button.

This means allowed_values is always required. Default
values are not required, but will be respected if
provided.

It also means only single values are accepted by the
radio filter (no arrays, just one string).