Add ContentItem model
The ContentItem model will be used for fetching and representing
content store items, and providing development mode local JSONs.

The intent behind adding this is to remove content item fetch /
representation logic from FinderApi. This will make both classes
a bit simpler by separating the concerns. In a subsequent commit
I will switch finder_api to recieve a content_item rather than
a base path.

The end goal is to make it easier to see how a GET request to
/:finder-path is handled.