Use ContentItem model to fetch content items
This removes content item fetch logic from FinderApi.
It makes use of the new ContentItem model. Ideally this
will make FinderApi a bit simpler, and make it easier
to simplify that class further.

In a later commit I would like to refactor the way that
content items and params translate into rummager queries
and presented data to the views. At the moment we send
data to FinderApi and AdvancedFinderApi which manipulates
the content item hash to present results, which could be
made a bit simpler.