Refactor api to presenter data flow
This removes augmentation code (where the content item hash is
manipulated by FinderApi and AdvancedSeaerchFinderApi) and
adds the same functionality to the presenters of those finders.

The intent is to remove the manipulation of the content item hash.
The augmentation of content item can be hard to follow, especially
considering that in the case of advanced search the same hash is
being manipulated across multiple classes before being handed to a
presenter.

The next step will be to remove the remaining augmentation code,
in place of a clearer data flow. Currently the way facet allowed_values
are added to the content_item hash is sub optimal and could be
moved from the FinderApi into its own class (FacetDataPresenter?)
in a future commit.