Add top result in documents For top result to be true, it must first meet the following criteria: - Must be more than 2 documents - Must have sort option key '-relevance' - Must be the business finder
Then the 2 most relevant documents are compared to their es_score's and if
there is a significant difference in score, then it is given a top result of true.

We need to add the description to the top result, therefore needed it publicly available to use in ResultSetPresenter.

We found the average score on the top 500 searches and found that 7 was the most suitable number.