Fix facet tag removal for multiple search terms
This commit fixes a bug with facet tag removal for text searches. Facet tag removal was always removing the first instance of that keyword, even if it was part of another word, e.g: "therefore the". In this case, removing "the" would result in "refore the". This commit fixes that.