Remove email subscription query validation logic
Currently we make a query to rummager to check whether the keys and
values are acceptible before we create a subscriber list with those
values.

This is not needed. We should validate the params on email-alert-api
end, where the values are consumed. Additionally, this particular
validation logic was added for the reject_content_purpose_supergroup
feature which has been removed.