Adds in lib object for publications routes
Adds a big list of special behaviour for the publications redirections.
Depending on the publication_filter_option that's selected, there's a
variety of redirections, which this addresses.

Tests are also included for each of these special redirects, as well as
the default path to search/all when nothing special shoul happen.