Add support for filter_values to email_signup facet_choices
This will enable us to send an array of values to the email alert api
when given a single key as a facet_choice.

In this case I have used this for the Research and statistics
Document types facet.

This is different to the option_lookup pattern. Where option_lookup
provides a similar behaviour, this is a bit clearer as the object
containing filter_values also contains the key for that value.