Add AB test to show answer box in business finder
This wraps the answer box that is displayed if the user does a keyword
search that returns a "best bet" in an A/B test.

The answer box is only displayed on the business readiness finder. As
the ResultSetPresenter is already handling that part of the logic, the
controller only determines which control group the user is in and
whether they "could" see the answer box, if they were on the correct
finder.