Fix for external links
I've not found any evidence of internal links that don't start with a '/'
so I'm not sure why we're doing this check.

It's used in `URI.join(Plek.current.website_root, path)` and was added in
https://github.com/alphagov/finder-frontend/commit/a408d1436707b960c86c2f26d3ca5d78615eb9ea

It's used in `absolute_url_for` in application_helper.rb, but that code treats
the path the same whether it has a leading slash or not:

```
irb(main):002:0> URI.join("https://www.gov.uk", "bernard")
=> #<URI::HTTPS https://www.gov.uk/bernard>
irb(main):003:0> URI.join("https://www.gov.uk", "/bernard")
=> #<URI::HTTPS https://www.gov.uk/bernard>
```

The test fixture seems to indicate that it might have been specialist content
from back in the day that was missing the leading '/', but this looks to have
been fixed.