Update feature test for answer box
We need get around the fact that the answer box only displays for
variant "B" of FinderAnswerABTest.

We’ve had to do a bit of fiddling to get this to work. Ideally we would
have updated the header and set the variant to “B” there, but Selenium
doesn’t allow that. That’s why the govuk_ab_testing gem is being stubbed
instead.

We also find out that “filling in” keywords doesn’t actually submit the form,
and the results aren’t being returned without it.

We’re only testing the “B” variant, because with javascript enabled, the
stubs aren’t reloaded between tests, so only one test will pass, either for
the “A” variant, or the “B”, but not both.

The mechanics of the A/B test are being tested in the controller and presenter
tests so it’s ok to only test that the top result is displaying when expected.
If there was more time, it would be worth trying to get both variants working
for completeness.