Add additional custom dimension for relevant result
The new custom dimension is Relevant result shown (83). This should be set to 'Yes' if a  search in the finder search box results in a 'relevant result'. (The null label is 'No'.) This is set at a hit/pageview level, since a user may do several searches in their session.

Ticket:  https://trello.com/c/lwQK2Pi7/336-set-up-a-b-test-for-answer-boxes