Remove updated options from sort options when statistics upcoming is selected
This commit removes the options `Updated (newest)` and `Updated (oldest)` from the sort options when the user chooses to view `Statistics (upcoming)`. If the user later chooses to view `Statistics (published)` or `Research`, the options are added back.

In addition, the current default sort option is `Updated (newest)`; this presents us a problem after removing it for `Statistics (upcoming)` if we need to default back to showing it, either because:
- the user previously had a keyword search with a sort option of `Relevancy`, and then removed that keyword search, or
- the user previously sorted by `Upcoming (newest)` or `Upcoming (oldest)` and then chose to view `Statictics (upcoming)`

If for either of the above reasons, we need to default to an option, we will default to `Release date (latest)`.