Limit AB test header to the page in test
This header tells fastly that it needs to cache a version of the page for
each value of the header.  This is only the case for the business finder
for this test (or so I believe).

This should reduce the number of requests that hit origin for other pages in
finder-frontend.