Truncate long queries, to match search-api
search-api rejects queries which are longer than 512 characters, but
we do get some legitibale (ie, human-looking) queries over this
length.  These queries are generally long questions, asked as if you
were talking to a human (not performing a search).  Searching for such
a thing is very unlikely to find good results, there's too much noise
in the keywords, but giving an error page isn't a good experience.

So we're going to silently truncate the query to match the search-api
limits, and do a search (which will still return unhelpful results,
for the same reason as previously).  At somepoint we'll do some UI
work to direct people to the contact form if they enter a long query.